==> 1_DataCollection.ipynb <==
## Import and data urls
We will load in three tables; title_basics, titles_akas, and title_ratings

import pandas as pd
import numpy as np

url1 = "https://datasets.imdbws.com/title.basics.tsv.gz"
url2 = "https://datasets.imdbws.com/title.akas.tsv.gz"
url3 = "https://datasets.imdbws.com/title.ratings.tsv.gz"

## Clean akas dataframe

%%time

# Load akas Data
# Because our data is seperated by a tab we indicate set='\t'.
# We will also use low_memory=False because the tables are very large
akas = pd.read_csv(url2, sep='\t', low_memory=False)

%%time

# Convert \\n to null (np.nan)
akas.replace({'\\N':np.nan}, inplace=True)

# Keep only the US movies
akas = akas[(akas['region'] == 'US')]
akas.info()

## Clean title.basics dataframe

%%time

# Load basics Data
basics = pd.read_csv(url1, sep='\t', low_memory=False)

# Replace '\\N' with np.nan
basics.replace({'\\N':np.nan}, inplace=True)

# Drop any rows that have null values in either genre or runtimeMinutes.
basics = basics.dropna(subset=['runtimeMinutes', 'genres'])

# Keep only titleType==Movie
is_Movie = basics['titleType'] == 'movie'
basics = basics[is_Movie]

# Drop all rows where the start year is not between 2000-2022.

# We first need to change the column type to a numeric type.
print("\nData Type before conversion : ", basics['startYear'].dtype)
basics['startYear'] = pd.to_numeric(basics['startYear'])
print("\nData Type after conversion : ", basics['startYear'].dtype)

#Now we can make and apply our filters
filter2000 = basics['startYear'] >= 2000 
filter2022 = basics['startYear'] <=2022
basics = basics[filter2000 & filter2022]

# Exclude movies that are included in the documentary category.
is_documentary = basics['genres'].str.contains('documentary', case=False)
basics = basics[~is_documentary]

# Filter the basics table down to only include the US movies by using the filtered akas dataframe above
keepers_basics = basics['tconst'].isin(akas['titleId'])
basics = basics[keepers_basics]

basics.info()

## Clean ratings dataframe

%%time

# Load ratings Data
ratings = pd.read_csv(url3, sep='\t', low_memory=False)

# Replace \\n with null
ratings.replace({'\\N':np.nan}, inplace=True)

# Filter the ratings table down to only include the US movies by using the filtered akas dataframe
keepers = ratings['tconst'].isin(akas['titleId'])
ratings = ratings[keepers]

ratings.info()

## save cleaned files in Data folder

# Save current dataframes to files.
basics.to_csv("Data/title_basics.csv.gz",compression='gzip',index=False)
akas.to_csv("Data/title_akas.csv.gz",compression='gzip',index=False)
ratings.to_csv("Data/title_ratings.csv.gz",compression='gzip',index=False)
print("Done!")
