==> 2_Extract from TMDB.ipynb <==
# Extracting the budget, revenue, and MPAA-Rating (G/PG/PG-13/R).
MPAA-Rating is called 'Certification'.

## import necessary libraries 

# Regular imports
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
# Special imports
import tmdbsimple as tmdb
import os, json, math, time
from tqdm.notebook import tqdm_notebook

with open('/Users/default.DESKTOP-0LDO0LD/.secret/tmdb.json', 'r') as f:
    login = json.load(f)
## Display the keys of the loaded dict
login.keys()

tmdb.API_KEY =  login['api-key']
print("Done!")

# show the new files in the current folder.
FOLDER = "Data/"
os.makedirs(FOLDER, exist_ok=True)
os.listdir(FOLDER)

# Load in the dataframe from project part 1 as basics:
basics = pd.read_csv('Data/title_basics.csv.gz')

YEARS_TO_GET = list(range(2003, 2020)) # speciftt the years we want to extract.
errors = [] # save the id's and error messages for any movie that causes an error. 

# Appends a list of records (new_data) to a json file (filename). 
# Adapted from: https://www.geeksforgeeks.org/append-to-json-file-using-python/
   
def write_json(new_data, filename):     
    with open(filename,'r+') as file:
        # First we load existing data into a dict.
        file_data = json.load(file)
        
        ## Choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
             file_data.append(new_data)
                
        # Sets file's current position at offset.
        file.seek(0)
        
        # convert back to json.
        json.dump(file_data, file)

# This function will add the certification (MPGG Rating)
def get_movie_with_rating(movie_id):
    # Get the movie object for the current id
    movie = tmdb.Movies(movie_id)
    
    # save the .info .releases dictionaries
    info = movie.info()
    releases = movie.releases()
    
    # Loop through countries in releases
    for c in releases['countries']:
        # if the country abbreviation==US
        if c['iso_3166_1' ] =='US':
            ## save a "certification" key in the info dict with the certification
           info['certification'] = c['certification']
    return info

%%time
# Start of OUTER loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc='YEARS', position=0):
    #Defining the JSON file to store results for year
    JSON_FILE = f'{FOLDER}tmdb_api_results_{YEAR}.json'
    # Check if file exists
    file_exists = os.path.isfile(JSON_FILE)

    # If it does not exist: create it
    if file_exists == False:
    # save an empty dict with just "imdb_id" to the new json file.
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)

    #Saving new year as the current df
    df = basics.loc[basics['startYear']==YEAR].copy()
    # saving movie ids to list
    movie_ids = df['tconst'].copy()
    
    # Load existing data from json into a dataframe called "previous_df"
    previous_df = pd.read_json(JSON_FILE)
    
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]

    #Get index and movie id from list
    # INNER Loop
    for movie_id in tqdm_notebook(movie_ids_to_get, desc=f'Movies from {YEAR}', position=1, leave=True):
        try:
            # Retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)  
            # Append/extend results to existing file using a pre-made function
            write_json(temp,JSON_FILE)
            # Short 20 ms sleep to prevent overwhelming server
            time.sleep(0.02)
            
        except Exception as e:             
            errors.append([movie_id, e])

    # Save files
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f"{FOLDER}final_tmdb_data_{YEAR}.csv.gz", compression="gzip", index=False)

print(f"- Total errors: {len(errors)}")

# Lets just see if everything worked for the first two years
df2000 = pd.read_csv("Data/final_tmdb_data_2000.csv.gz")
df2001 = pd.read_csv("Data/final_tmdb_data_2001.csv.gz")

# Preview data for 2000
df2000.head()

# Preview data for 2001
df2001.head()

# Lets check whats in our Data folder.
os.listdir(FOLDER)
