==> part 3.ipynb <==
# Regular imports
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
# Special imports
import tmdbsimple as tmdb
import os, json, math, time
from tqdm.notebook import tqdm_notebook

import pymysql
pymysql.install_as_MySQLdb()
from sqlalchemy import create_engine
from sqlalchemy_utils import create_database, database_exists

df_basics = pd.read_csv('Data/title_basics.csv.gz')

# Discard any unnecessary columns
df_basics.drop(columns=['originalTitle', 'isAdult', 'titleType'], inplace=True)
df_basics.head()

# Create database 
with open('/Users/default.DESKTOP-0LDO0LD/.secret/mySQL.json') as f:
    login = json.load(f)
    
connection_str = f"mysql+pymysql://root:{login['password']}@localhost/movies"
engine = create_engine(connection_str)
if database_exists(connection_str) == False:
    create_database(connection_str)
else: 
    print("Already there my friend")

# Lets see what is stored in the genres columns and check what type it is.
temp = df_basics.loc[0, 'genres'].split(',')
print(type(temp))
temp

# Lets create
df_basics['genres_split'] = df_basics['genres'].str.split(',')
temp = df_basics.loc[0, 'genres'].split(',')
df_basics.head()

# Make a list of genres
exploded = df_basics.explode('genres_split')
exploded

df_title_genre = exploded[['tconst', 'genres_split']]
df_title_genre

genres_to_make = sorted(exploded['genres_split'].dropna().unique())
genres_to_make

## Making the genre mapper dictionary
genre_ints = range(len(genres_to_make))
genre_map = dict(zip(genres_to_make, genre_ints))
genre_map

# make genre map a data frame
genre_lookup = pd.DataFrame({'Genre_Name': genre_map.keys(), 'Genre_Id': genre_map.values()})
genre_lookup

## make new integer genre_id and drop string genres
df_title_genre['genre_id'] = df_title_genre['genres_split'].map(genre_map)
df_title_genre = df_title_genre.drop(columns='genres_split')
df_title_genre 

# Delete the genres and genres_split columns
df_basics.drop(columns=['genres','genres_split'], inplace=True)

# get max string length
max_str_len_tconst = df_basics['tconst'].fillna('').map(len).max()
max_str_len_title = df_basics['primaryTitle'].fillna('').map(len).max()

from sqlalchemy.types import *
## Calculate max string lengths for object columns
# already in line 2 and 3
# key_len = basics['tconst'].fillna('').map(len).max()
# title_len = basics['primaryTitle'].fillna('').map(len).max()

## Create a schema dictonary using Sqlalchemy datatype objects
df_schema = {
    "tconst": String(max_str_len_tconst+1), 
    "primaryTitle": Text(max_str_len_title+1),
    'startYear':Float(),
    'endYear':Float(),
    'runtimeMinutes':Integer()}

# Save to sql with dtype and index=False
df_basics.to_sql('title_basics', engine, dtype=df_schema, if_exists='replace', index=False)

engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);');

pd.read_sql("Select * from title_basics", engine)

# Load the results from previous step
df_combined = pd.read_csv('Data/tmdb_results_combined.csv.gz')

# drop duplicates 
df_combined.drop_duplicates(inplace=True)
print("The number of duplicates : ", df_combined.duplicated().sum())

#  remove rows that are imdb_id  = 0
df_combined = df_combined[df_combined['imdb_id'] != '0']

df_combined.head()

tmdb_schema = {"imdb_id": String(max_str_len_tconst+1)}

# Add tmdb_data to database. the primary key is imdb_id
df_combined.to_sql('tmdb_data', engine, dtype=tmdb_schema, if_exists='replace', index=False)
engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);');

# Test query
pd.read_sql("Select * from tmdb_data", engine)

# Add title_genre todatabase. There is no primary key.
df_title_genre.to_sql('title_genres', engine, if_exists='replace', index=False)

# Test query
pd.read_sql("Select * from title_genres", engine)

ratings_df = pd.read_csv('Data/title_ratings.csv.gz')
ratings_df.head()

# Add title_ratings to database. The primary key is tconst
ratings_df.to_sql('title_ratings', engine, dtype=df_schema, if_exists='replace', index=False)
engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);');

# Test Query
pd.read_sql("Select * from title_ratings LIMIT 5", engine)

# Add genre look up table to database. Genre_Id is the primary key.
genre_lookup.set_index('Genre_Id').to_sql('genres', engine, if_exists='replace')

# Test Query
pd.read_sql("Select * from genres", engine)

By this time you should have the following tables in the database:
 - title_basics
 - title_ratings
 - title_genres 
 - genres 
 - tmdb_data
 
 
 
Make sure to set a Primary Key for each table that isn't a joiner table (e.g. title_genres is a joiner table).

After creating each table, show the first 5 rows of that table using a SQL query.

Make sure to run the "SHOW TABLES" SQL query at the end of your notebook to show that all required tables have been created.

pd.read_sql("SHOW TABLES", engine)
